[timeseries] Create an abstraction layer communicating  with the timeseries #65
