+++ openwisp_monitoring/device/base/models.py
@@ -19,7 +19,7 @@ from swapper import load_model
 
 from openwisp_utils.base import TimeStampedEditableModel
 
-from ...db import TimeseriesDB, device_data_query
+from ...db import device_data_query, timeseries_db
 from ...monitoring.signals import threshold_crossed
 from .. import settings as app_settings
 from ..schema import schema
@@ -91,12 +91,12 @@ class AbstractDeviceData(object):
     @property
     def data(self):
         """
-        retrieves last data snapshot from InfluxDB
+        retrieves last data snapshot from Timeseries Database
         """
         if self.__data:
             return self.__data
         q = device_data_query.format(SHORT_RP, self.__key, self.pk)
-        points = list(TimeseriesDB.query(q).get_points())
+        points = timeseries_db.get_list_query(q, precision=None)
         if not points:
             return None
         self.data_timestamp = points[0]['time']
@@ -161,14 +161,14 @@ class AbstractDeviceData(object):
 
     def save_data(self, time=None):
         """
-        validates and saves data to InfluxDB
+        validates and saves data to Timeseries Database
         """
         self.validate_data()
         if app_settings.MAC_VENDOR_DETECTION:
             self.add_mac_vendor_info()
         # TODO: Rename the parameters, since they might be called
         # differently in the other database (eg: tags/labels)
-        TimeseriesDB.write(
+        timeseries_db.write(
             name=self.__key,
             values={'data': self.json()},
             tags={'pk': self.pk},
