+++ openwisp_monitoring/device/tests/test_models.py
@@ -9,10 +9,7 @@ from swapper import load_model
 from openwisp_controller.connection.models import Credentials, DeviceConnection
 from openwisp_utils.tests import catch_signal
 
-from ...db import TimeseriesDB
-from .. import settings as app_settings
 from ..signals import health_status_changed
-from ..utils import SHORT_RP
 from . import DeviceMonitoringTestCase
 
 Check = load_model('check', 'Check')
@@ -304,14 +301,6 @@ class TestDeviceData(BaseTestCase):
         self.assertIsNone(dd.data)
         dd = DeviceData(data=self._sample_data)
         self.assertEqual(dd.data, self._sample_data)
-
-    def test_retention_policy(self):
-        rp = TimeseriesDB.get_db().get_list_retention_policies()
-        self.assertEqual(len(rp), 2)
-        self.assertEqual(rp[1]['name'], SHORT_RP)
-        self.assertEqual(rp[1]['default'], False)
-        duration = app_settings.SHORT_RETENTION_POLICY
-        self.assertEqual(rp[1]['duration'], duration)
 
     def test_device_deleted(self):
         d = self._create_device()
