+++ openwisp_monitoring/monitoring/tests/test_models.py
@@ -9,7 +9,6 @@ from swapper import load_model
 
 from openwisp_utils.tests import catch_signal
 
-from ...db import TimeseriesDB
 from ..signals import post_metric_write, pre_metric_write, threshold_crossed
 from . import TestMonitoringMixin
 
@@ -106,82 +105,6 @@ class TestModels(TestMonitoringMixin, TestCase):
         self.assertEqual(m.id, m2.id)
         self.assertEqual(m2.name, m.name)
         self.assertFalse(created)
-
-    def test_write(self):
-        TimeseriesDB.write('test_write', dict(value=2), database=self.TEST_DB)
-        measurement = list(
-            TimeseriesDB.query(
-                'select * from test_write', database=self.TEST_DB
-            ).get_points()
-        )[0]
-        self.assertEqual(measurement['value'], 2)
-
-    def test_general_write(self):
-        m = self._create_general_metric(name='Sync test')
-        m.write(1)
-        measurement = list(TimeseriesDB.query('select * from sync_test').get_points())[
-            0
-        ]
-        self.assertEqual(measurement['value'], 1)
-
-    def test_object_write(self):
-        om = self._create_object_metric()
-        om.write(3)
-        content_type = '.'.join(om.content_type.natural_key())
-        q = (
-            "select * from test_metric WHERE object_id = '{0}'"
-            " AND content_type = '{1}'".format(om.object_id, content_type)
-        )
-        measurement = list(TimeseriesDB.query(q).get_points())[0]
-        self.assertEqual(measurement['value'], 3)
-
-    def test_general_same_key_different_fields(self):
-        down = self._create_general_metric(
-            name='traffic (download)', key='traffic', field_name='download'
-        )
-        down.write(200)
-        up = self._create_general_metric(
-            name='traffic (upload)', key='traffic', field_name='upload'
-        )
-        up.write(100)
-        measurement = list(
-            TimeseriesDB.query('select download from traffic').get_points()
-        )[0]
-        self.assertEqual(measurement['download'], 200)
-        measurement = list(
-            TimeseriesDB.query('select upload from traffic').get_points()
-        )[0]
-        self.assertEqual(measurement['upload'], 100)
-
-    def test_object_same_key_different_fields(self):
-        user = self._create_user()
-        user_down = self._create_object_metric(
-            name='traffic (download)',
-            key='traffic',
-            field_name='download',
-            content_object=user,
-        )
-        user_down.write(200)
-        user_up = self._create_object_metric(
-            name='traffic (upload)',
-            key='traffic',
-            field_name='upload',
-            content_object=user,
-        )
-        user_up.write(100)
-        content_type = '.'.join(user_down.content_type.natural_key())
-        q = (
-            "select download from traffic WHERE object_id = '{0}'"
-            " AND content_type = '{1}'".format(user_down.object_id, content_type)
-        )
-        measurement = list(TimeseriesDB.query(q).get_points())[0]
-        self.assertEqual(measurement['download'], 200)
-        q = (
-            "select upload from traffic WHERE object_id = '{0}'"
-            " AND content_type = '{1}'".format(user_up.object_id, content_type)
-        )
-        measurement = list(TimeseriesDB.query(q).get_points())[0]
-        self.assertEqual(measurement['upload'], 100)
 
     def test_read_general_metric(self):
         m = self._create_general_metric(name='load')
