+++ tests/openwisp2/settings.py
@@ -1,4 +1,3 @@
-import importlib
 import os
 import sys
 from datetime import timedelta
